move dicehub to core in proto
